add authentcation middleware and update all router files.

+++ routes/bugs.js
@@ -1,38 +1,53 @@
 const { Bug, validateBug } = require("../models/bug");
+const auth = require("../middleware/auth");
 const express = require("express");
 const router = express.Router();
 
 // NOTE: get all Bugs
-router.get("/", async (req, res) => {
-  let Bugs;
-  // INFO: get all Bugs
-  Bugs = await Bug.find().populate("user", "-isAdmin");
+router.get("/", auth, async (req, res) => {
+  let bugs;
+  // INFO: admin will get all bugs
+  if (req.user.isAdmin) {
+    bugs = await Bug.find().populate("user", "-isAdmin").select("-__v");
+    return res.send(bugs);
+  }
 
-  res.send(Bugs);
+  // INFO: user get all bugs
+  bugs = await Bug.find({ user: req.user._id }).populate("user", "-isAdmin");
+
+  res.send(bugs);
 });
 
-// NOTE: add new Bug
-router.post("/", async (req, res) => {
+// NOTE: add new bug
+router.post("/", auth, async (req, res) => {
   req.body.user = req.user._id;
   // INFO:  validate data send by user
   const { error } = validateBug(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
-  const Bug = new Bug({
+  const bug = new Bug({
     title: req.body.title,
     description: req.body.description,
     status: req.body.status,
     projectId: req.body.projectId,
     user: req.user._id,
   });
-  await Bug.save();
+  await bug.save();
 
-  res.status(201).send(Bug);
+  res.status(201).send(bug);
 });
 
-// NOTE: update Bug
-router.patch("/:id", async (req, res) => {
-  const Bug = await Bug.findByIdAndUpdate(
+// NOTE: update bug
+router.patch("/:id", auth, async (req, res) => {
+  let bug = await Bug.findById(req.params.id);
+  if (!bug)
+    return res.status(404).send(" The bug with given ID was not found.");
+
+  // INFO: the owner or admin can update the bug
+  if (req.user._id !== bug.user._id || req.user.isAdmin === "false") {
+    return res.status(405).send("Method not allowed.");
+  }
+  bug = await Bug.findByIdAndUpdate(
     req.params.id,
     {
       title: req.body.title,
@@ -44,30 +59,39 @@ router.patch("/:id", async (req, res) => {
     { new: true }
   );
 
-  if (!Bug)
-    return res.status(404).send(" The Bug with given ID was not found.");
-  res.send(Bug);
+  res.send(bug);
 });
 
-// NOTE: delete one Bug by id
-router.delete("/:id", async (req, res) => {
-  const Bug = await Bug.findByIdAndRemove(req.params.id);
-  if (!Bug)
-    return res.status(404).send(" The Bug with given ID was not found.");
+// NOTE: delete one bug by id
+router.delete("/:id", auth, async (req, res) => {
+  let bug = await Bug.findById(req.params.id);
+  if (!bug)
+    return res.status(404).send(" The bug with given ID was not found.");
 
-  return res.send(Bug);
+  // INFO: the owner or admin can delete the bug
+  if (req.user._id !== bug.user._id || req.user.isAdmin === "false") {
+    return res.status(405).send("Method not allowed.");
+  }
+  bug = await Bug.findByIdAndRemove(req.params.id);
+  return res.send(bug);
 });
 
-// NOTE: get one Bug route
+// NOTE: get one bug route
 router.get("/:id", async (req, res) => {
-  const Bug = await Bug.findById(req.params.id).populate(
+  let bug = await Bug.findById(req.params.id);
+  if (!bug)
+    return res.status(404).send(" The bug with given ID was not found.");
+
+  // INFO: the owner or admin can get the bug
+  if (req.user._id !== bug.user._id || req.user.isAdmin === "false") {
+    return res.status(405).send("Method not allowed.");
+  }
+  bug = await Bug.findById(req.params.id).populate(
     "user",
     "name _id profileImage"
   );
-  if (!Bug)
-    return res.status(404).send(" The Bug with given ID was not found.");
 
-  res.send(Bug);
+  res.send(bug);
 });
 
 module.exports = router;
